% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsimulate2bhru.R
\name{gsim2bhru}
\alias{gsim2bhru}
\title{convert a BaseFile and MixFile output to long(-ish) format usable by BHRU}
\usage{
gsim2bhru(Num, coll2RU = NULL)
}
\arguments{
\item{Num}{the number of the data set to read in.  Assumes that things are
named like BaseFile_X.txt and MixFile_X.txt.  Num should be what you want
passed in for X there...}

\item{coll2RU}{a data frame with columns "repunit" and "collection".
Each collection is included once, and the "repunit" of the same row is the
reporting unit to which it belongs}
}
\description{
More later
}

