% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_paths.R
\name{ms2geno_binary}
\alias{ms2geno_binary}
\title{file path to be used in a call to ms2geno}
\usage{
ms2geno_binary()
}
\description{
This version checks to make sure it is there and throws an
error with a suggestion of how to get it if it is not there.
}

