% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_paths.R
\name{ms2geno_binary_path}
\alias{ms2geno_binary_path}
\title{return the path where ms2geno should be in the R system paths}
\usage{
ms2geno_binary_path()
}
\description{
return the path where ms2geno should be in the R system paths
}
