% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mig_mats.R
\name{sh_island_mig_mat}
\alias{sh_island_mig_mat}
\title{return the migration matrix for a simple hierarchical island model}
\usage{
sh_island_mig_mat(ru, Min, Mbt)
}
\arguments{
\item{ru}{a vector giving the numbers of collections within each
group.  For example, if there are 10 collections total and the first
two are in one group, the next three are in another and the final five
are in a third, then you would want \code{ru = c(2, 3, 5)}.}

\item{Min}{the M = 4Nem value desired within groups.  This will get
scaled by the number of collections within the group (Min / (N - 1))
where N is the number of collections in the group.}

\item{Mbt}{the M = 4Nem value desired between collections that are in
different groups. This will not be scaled by the number of collections not in
the group.}
}
\value{
Returns that matrix in text format to pass into ms
}
\description{
This provides the migration matrix for one of the simplest possible
island models with migration rates within and between groups.
}
